+++ end_to_end_tests/golden-record/my_test_api_client/api/tests/get_user_list.py
@@ -19,6 +19,7 @@ def _get_kwargs(
     url = "{}/tests/".format(client.base_url)
 
     headers: Dict[str, Any] = client.get_headers()
+    cookies: Dict[str, Any] = client.get_cookies()
 
     json_an_enum_value = []
     for an_enum_value_item_data in an_enum_value:
@@ -40,7 +41,7 @@ def _get_kwargs(
     return {
         "url": url,
         "headers": headers,
-        "cookies": client.get_cookies(),
+        "cookies": cookies,
         "timeout": client.get_timeout(),
         "params": params,
     }