+++ end_to_end_tests/golden-record/my_test_api_client/api/tests/optional_value_tests_optional_query_param.py
@@ -15,6 +15,7 @@ def _get_kwargs(
     url = "{}/tests/optional_query_param/".format(client.base_url)
 
     headers: Dict[str, Any] = client.get_headers()
+    cookies: Dict[str, Any] = client.get_cookies()
 
     json_query_param: Union[Unset, List[Any]] = UNSET
     if not isinstance(query_param, Unset):
@@ -28,7 +29,7 @@ def _get_kwargs(
     return {
         "url": url,
         "headers": headers,
-        "cookies": client.get_cookies(),
+        "cookies": cookies,
         "timeout": client.get_timeout(),
         "params": params,
     }